pyproject.toml
[tool.poetry]
name = "refractr"
version = "0.1.0"
description = ""
authors = ["André Honeiser <ahoneiser@mozilla.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "3.8.6"
doit = "~0.36.0"
jinja2 = "~3.1.4"
jsonschema = "3.2.0"
leatherman = "0.1.10"
nginx-config-builder = "1.0.1"
pytest = "7.1.2"
python-decouple = "3.3"
ruamel-yaml = "~0.18.6"
tldextract = "3.0.0"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
